fix: auth check component change
