+++ frontend/app/components/header/HeaderLogoutComponent.tsx
@@ -1,5 +1,3 @@
-
-
 const HeaderLogoutComponent = () => {
   return (
     <div>HeaderLogoutComponent</div>
